requirements.txt
discord.py==1.5.1
googletrans==3.0.0
numpy==1.19.5
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
webcolors==1.11.1
